Pluggable DAQ support, parts 1&2: Plumbing
Changeset reflects ModelManager introduction (parts 1 & 2), with just a
few additional changes at the GUI level to hold things together

+++ src/drivers/Daq.cpp
@@ -1,6 +1,6 @@
 #include "Daq.h"
 
-DAQ::DAQ(int devID) :
+DAQ::DAQ(size_t devID) :
     devID(devID),
     t(DAQClock.t)
 {
